requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pandas==2.2.2
python-dotenv==1.0.1
apache-airflow==2.9.3
apache-airflow-providers-postgres==5.11.0
apache-airflow-providers-http==4.12.0
